* sn2princ.c, hst_realm.c, an_to_ln.c: Cast argument to 	isupper()/isspace()/islower() to int.
	* t_std_conf.c (test_locate_kdc): krb5_locate_kdc expects a struct
	sockaddr *** as a third argument instead of sockaddr **.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13339 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/os/an_to_ln.c
@@ -388,7 +388,7 @@ aname_replacer(string, contextp, result)
 	 */
 	for (cp = *contextp; *cp; ) {
 	    /* Skip leading whitespace */
-	    while (isspace(*cp))
+	    while (isspace((int) (*cp)))
 		cp++;
 
 	    /*
@@ -520,7 +520,7 @@ rule_an_to_ln(context, rule, aname, lnsize, lname)
 				    *cout = '\0';
 				    current++;
 				    /* Point past number */
-				    while (isdigit(*current))
+				    while (isdigit((int) (*current)))
 					current++;
 				}
 				else
@@ -703,7 +703,7 @@ krb5_aname_to_localname(context, aname, lnsize_in, lname)
 		    /* Trim the value. */
 		    cp = &mapping_values[nvalid-1]
 			[strlen(mapping_values[nvalid-1])];
-		    while (isspace(*cp)) cp--;
+		    while (isspace((int) (*cp))) cp--;
 		    cp++;
 		    *cp = '\0';
