+++ src/lib/krb5/os/sn2princ.c
@@ -99,7 +99,7 @@ krb5_sname_to_principal(context, hostname, sname, type, ret_princ)
 
 	if (type == KRB5_NT_SRV_HST)
 	    for (cp = remote_host; *cp; cp++)
-		if (isupper(*cp))
+		if (isupper((int) (*cp)))
 		    *cp = tolower(*cp);
 
 	/*
